// typescript/user-interface/packages/weavess-core/src/ts/types/types.ts
export enum DisplayType {
  /** String representation of line type 'LINE' */
  LINE = 'LINE',

  /** String representation of line type 'SCATTER' */
  SCATTER = 'SCATTER'
}
// ...
export interface TimeRange {
  /** Start Time in seconds */
  startTimeSecs: number;

  /** End Time in seconds */
  endTimeSecs: number;
}
// ...
export interface ChannelSegmentBoundaries {
  /** Maximum value of top */
  topMax: number;

  /** Maximum value of bottom */
  bottomMax: number;

  /** Average of channel */
  channelAvg: number;

  /** The maximum magnitude of this channel */
  offset: number;

  /** Channel segment id */
  channelSegmentId: string;

  /** Samples count used to compute average helps
      when merging multiple boundaries */
  samplesCount?: number;

  /** topMax epoch seconds found */
  topMaxSecs?: number;

  /** bottomMax epoch seconds found */
  bottomMaxSecs?: number;
}
// ...
export interface ChannelSegment {
  /** Name of the configured input of the channel, this will be stable between filters and can be used
   * to select original (unfiltered) channel record */
  configuredInputName: string;

  /** Channel Name of the Waveform */
  channelName: string;

  /** Waveform filter of the Waveform the default (Raw) data is 'unfiltered' */
  wfFilterId: string;

  /** Flags this channel segment is selected
   * when corresponding SD is selected */
  isSelected: boolean;

  /** Collection of data segments */
  dataSegments: DataSegment[];

  /** Channel segment description */
  description?: string;

  /** Color of the label for the description */
  descriptionLabelColor?: string;

  /** processed Amplitude values during Typed Array conversion */
  channelSegmentBoundaries?: ChannelSegmentBoundaries;

  /** Units of the channel segment time series data */
  units?: string;

  /** The type of time series in this segment. */
  timeseriesType?: string;
}

/**
 * Represents data values by sample start time and sample rate
 */
export interface DataBySampleRate {
  /** Epoch start time in seconds */
  startTimeSecs: number;

  /** Epoch end time in seconds */
  endTimeSecs: number;

  /** Sample Rate */
  sampleRate: number;

  /** Collection representing data segment data */
  values: Float32Array | number[];
}

export interface DataClaimCheck extends Omit<DataBySampleRate, 'values'> {
  /** Collection representing data segment data */
  values: Float32Array | number[] | undefined;

  /** Claim check ID. Used to fetch and perform operations on this waveform sample data. */
  id: string;

  /** Used in conversion of waveform to set x (time) position */
  domainTimeRange: TimeRange;
}

/**
 * Represents a time value pair.
 */
export interface TimeValuePair {
  timeSecs: number;
  value: number;
}

/**
 * Represents data values by time.
 */
export interface DataByTime {
  values: TimeValuePair[] | Float32Array;
}

/** Data Segment */
export interface DataSegment {
  /** Color */
  color?: string;

  /** Display type */
  displayType?: DisplayType[];

  /** Point size */
  pointSize?: number;

  /** Collection representing data segment data */
  data: DataBySampleRate | DataByTime | DataClaimCheck;
}